third_party/WebKit/LayoutTests/http/tests/websocket/tests/hybi/invalid-subprotocols-expected.txt

CONSOLE MESSAGE: Wrong protocol for WebSocket 'a b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a,b'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'a\u0000b'
CONSOLE MESSAGE: Wrong protocol for WebSocket '\u3042'
CONSOLE MESSAGE: Wrong protocol for WebSocket 'c d'
CONSOLE MESSAGE: WebSocket protocols contain duplicates: 'a'
WebSocket should fail if invalid subprotocols are specified

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new WebSocket(url, 'a b') threw exception SyntaxError: An invalid or illegal string was specified..
PASS new WebSocket(url, 'a,b') threw exception SyntaxError: An invalid or illegal string was specified..
PASS new WebSocket(url, 'a\0b') threw exception SyntaxError: An invalid or illegal string was specified..
PASS new WebSocket(url, '\u3042') threw exception SyntaxError: An invalid or illegal string was specified..
PASS new WebSocket(url, ['a', 'b', 'c d']) threw exception SyntaxError: An invalid or illegal string was specified..
PASS new WebSocket(url, ['a', 'b', 'c', 'a']) threw exception SyntaxError: An invalid or illegal string was specified..
PASS successfullyParsed is true

TEST COMPLETE
